refactor: 자동차 경주(RacingCarGame) 결과 전달 시 dto 사용하도록 코드 수정
