+++ src/main/java/racingcar/step3/RacingCarGameApplication.java
@@ -3,12 +3,14 @@ package racingcar.step3;
 import racingcar.step3.domain.Car;
 import racingcar.step3.domain.CarFactory;
 import racingcar.step3.domain.RacingCarGame;
-import racingcar.step3.domain.RacingCarGameResult;
+import racingcar.step3.domain.RacingCarGameFactory;
+import racingcar.step3.dto.RacingCarGameResultDto;
 import racingcar.step3.ui.InputValue;
 import racingcar.step3.ui.InputView;
 import racingcar.step3.ui.ResultView;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class RacingCarGameApplication {
 
@@ -18,10 +20,9 @@ public class RacingCarGameApplication {
 
   public void run(String[] args) {
     InputValue input = InputView.receive();
-
     List<Car> cars = CarFactory.of(input.getNumberOfCars());
-    RacingCarGame racingCarGame = new RacingCarGame(input.getTryCount(), cars);
-    List<RacingCarGameResult> results = racingCarGame.start();
+    List<RacingCarGame> racingCarGames = RacingCarGameFactory.of(input.getTryCount(), cars);
+    List<RacingCarGameResultDto> results = racingCarGames.stream().map(RacingCarGame::start).collect(Collectors.toList());
 
     ResultView.printResult(results);
   }
